Events commands

+++ core/uniMatch/event/application/commands/DeleteEventCommand.ts
@@ -5,6 +5,8 @@ import { IEventRepository } from "../ports/IEventRepository";
 import { IEventBus } from "@/core/shared/application/IEventBus";
 import { IFileHandler } from "@/core/shared/application/IFileHandler";
 import { NotFoundError } from "@/core/shared/exceptions/NotFoundError";
+import { AuthorizationError } from "@/core/shared/exceptions/AuthorizationError";
+import { FileHandler } from "@/core/uniMatch/event/infrastructure/FileHandler";
 
 export class DeleteEventCommand implements ICommand<DeleteEventDTO, void> {
     private readonly repository: IEventRepository;
@@ -24,11 +26,13 @@ export class DeleteEventCommand implements ICommand<DeleteEventDTO, void> {
             const event = await this.repository.findById(request.eventId);
 
             if (!event) {
-                return Result.failure<void>(new NotFoundError("Event not found"));
+                return Result.failure<void>(new NotFoundError("Event"));
             }
-            
+            if( event.ownerId !== request.userId) {
+                return Result.failure<void>(new AuthorizationError("You are not authorized to delete this event"));
+            }
             if (event?.thumbnail) {
-                this.fileHandler.delete(event.thumbnail);
+                await this.fileHandler.delete(event.thumbnail);
             }
 
             event.delete();
